Third batch of review updates to ESP32 Azure IoT Kit sample
