+++ demos/projects/NXP/mimxrt1060/config/demo_config.h
@@ -7,6 +7,12 @@
 /* FreeRTOS config include. */
 #include "FreeRTOSConfig.h"
 
+/*
+ * This Model ID is tightly tied to the code implementation in `sample_azure_iot_pnp_simulated_device.c`
+ * If you intent to test a different Model ID, please provide the implementation of the model on your application.
+ */
+#define sampleazureiotMODEL_ID                                "dtmi:com:example:Thermostat;1"
+
 /**************************************************/
 /******* DO NOT CHANGE the following order ********/
 /**************************************************/
